cluster-image-builder/requirements.txt
llama_cpp_python==0.3.7
fastapi==0.109.2
typing_extensions==4.12.2
sse-starlette==2.1.3
starlette-context==0.3.6
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.8.1
huggingface-hub==0.29.1
ray[serve]
bentoml==1.4.6
PyYAML==6.0.1
vllm==0.7.3
kantoku==0.18.3
openai==1.65.2